Column specification for data about Completion at Private High Schools

Column specification for data about Completion at Private High Schools

Usage

SPEC_FOR_PRIVATE_HS_COMPLETION

Format

SPEC_FOR_PRIVATE_HS_COMPLETION

A list of 17 elements

SCHOOL_YEAR

<chr>

AGENCY_TYPE

<chr>

CESA

<chr>

COUNTY_CODE

<chr>

COUNTY

<chr>

CITY

<chr>

DISTRICT_CODE

<chr>

SCHOOL_CODE

<chr>

GRADE_GROUP

<chr>

CHOICE_IND

<chr>

DISTRICT_NAME

<chr>

SCHOOL_NAME

<chr>

LOW_GRADE

<chr>

HIGH_GRADE

<dbl>

GROUP_BY

<chr>

GROUP_BY_VALUE

<chr>

STUDENT_COUNT

<int>

Source

https://dpi.wi.gov/wisedash/download-files/type?field_wisedash_upload_type_value=hs-completion-private-school